% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directTransfer.R
\name{directTransfer_date_view_dmsServer}
\alias{directTransfer_date_view_dmsServer}
\title{直接调拨按日期查询}
\usage{
directTransfer_date_view_dmsServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
直接调拨按日期查询
}
\examples{
directTransfer_date_view_dmsServer()
}
