% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directTransfer.R
\name{directTransfer_fbillno_view_dmsServer}
\alias{directTransfer_fbillno_view_dmsServer}
\title{DMS直接调拨按单据编号查询}
\usage{
directTransfer_fbillno_view_dmsServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
DMS直接调拨按单据编号查询
}
\examples{
directTransfer_fbillno_view_dmsServer()
}
