% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directTransfer.R
\name{directTransfer_sync_erpServer}
\alias{directTransfer_sync_erpServer}
\title{直接调拨同步至ERP}
\usage{
directTransfer_sync_erpServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
直接调拨同步至ERP
}
\examples{
directTransfer_sync_erpServer()
}
