% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherInbound.R
\name{otherInbound_fbillno_view_dmsServer}
\alias{otherInbound_fbillno_view_dmsServer}
\title{DMS其他入库按单据编号查询}
\usage{
otherInbound_fbillno_view_dmsServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
DMS其他入库按单据编号查询
}
\examples{
otherInbound_fbillno_view_dmsServer()
}
