% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherInbound.R
\name{otherInbound_sync_erpServer}
\alias{otherInbound_sync_erpServer}
\title{其他入库同步至ERP}
\usage{
otherInbound_sync_erpServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
其他入库同步至ERP
}
\examples{
otherInbound_sync_erpServer()
}
