% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherOutbound.R
\name{otherOutbound_sync_dmsServer}
\alias{otherOutbound_sync_dmsServer}
\title{其他出库同步至数据中台}
\usage{
otherOutbound_sync_dmsServer(input, output, session, dms_token, wms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}

\item{wms_token}{}
}
\value{
返回值
}
\description{
其他出库同步至数据中台
}
\examples{
otherOutbound_sync_dmsServer()
}
