% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherOutbound.R
\name{otherOutbound_sync_erpServer}
\alias{otherOutbound_sync_erpServer}
\title{其他出库同步至ERP}
\usage{
otherOutbound_sync_erpServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
其他出库同步至ERP
}
\examples{
otherOutbound_sync_erpServer()
}
