% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prd_pickmtr.R
\name{prd_pickmtr_dateRange_view_dmsServer}
\alias{prd_pickmtr_dateRange_view_dmsServer}
\title{生产领料_倒扣线边库按日期范围查询}
\usage{
prd_pickmtr_dateRange_view_dmsServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
生产领料_倒扣线边库按日期范围查询
}
\examples{
prd_pickmtr_dateRange_view_dmsServer()
}
