% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prd_pickmtrl_waste.R
\name{prd_pickmtrl_waste_date_view_dmsServer}
\alias{prd_pickmtrl_waste_date_view_dmsServer}
\title{生产领料-废料入库按日期查询}
\usage{
prd_pickmtrl_waste_date_view_dmsServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
生产领料-废料入库按日期查询
}
\examples{
prd_pickmtrl_waste_date_view_dmsServer()
}
