% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prd_pickmtrl_waste.R
\name{prd_pickmtrl_waste_sync_dmsServer}
\alias{prd_pickmtrl_waste_sync_dmsServer}
\title{生产领料-废料入库同步至数据中台}
\usage{
prd_pickmtrl_waste_sync_dmsServer(input, output, session, dms_token, wms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}

\item{wms_token}{}
}
\value{
返回值
}
\description{
生产领料-废料入库同步至数据中台
}
\examples{
prd_pickmtrl_waste_sync_dmsServer()
}
