% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prd_pickmtrl_waste.R
\name{prd_pickmtrl_waste_sync_erpServer}
\alias{prd_pickmtrl_waste_sync_erpServer}
\title{生产领料-废料入库同步至ERP}
\usage{
prd_pickmtrl_waste_sync_erpServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
生产领料-废料入库同步至ERP
}
\examples{
prd_pickmtrl_waste_sync_erpServer()
}
