% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prd_pickmtrl_waste.R
\name{prd_pickmtrl_wasteServer}
\alias{prd_pickmtrl_wasteServer}
\title{处理逻辑}
\usage{
prd_pickmtrl_wasteServer(input, output, session, dms_token, wms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
prd_pickmtrl_wasteServer()
}
