% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchaseInbound.R
\name{purchaseInbound_sync_erpServer}
\alias{purchaseInbound_sync_erpServer}
\title{采购入库同步至ERP}
\usage{
purchaseInbound_sync_erpServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
采购入库同步至ERP
}
\examples{
purchaseInbound_sync_erpServer()
}
