% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchaseInbound.R
\name{purchaseInboundServer}
\alias{purchaseInboundServer}
\title{采购入库功能汇总}
\usage{
purchaseInboundServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
采购入库功能汇总
}
\examples{
purchaseInboundServer()
}
