% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchaseReturn.R
\name{purchaseReturn_sync_dmsServer}
\alias{purchaseReturn_sync_dmsServer}
\title{采购退料同步}
\usage{
purchaseReturn_sync_dmsServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
采购退料同步
}
\examples{
purchaseReturn_sync_dmsServer()
}
