% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpd_returnstock.R
\name{rpd_returnstock_sync_erpServer}
\alias{rpd_returnstock_sync_erpServer}
\title{生产退库同步至ERP}
\usage{
rpd_returnstock_sync_erpServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
生产退库同步至ERP
}
\examples{
rpd_returnstock_sync_erpServer()
}
