% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOut.R
\name{saleOut_fbillno_view_dmsServer}
\alias{saleOut_fbillno_view_dmsServer}
\title{DMS销售出库按单据编号查询}
\usage{
saleOut_fbillno_view_dmsServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
DMS销售出库按单据编号查询
}
\examples{
saleOut_fbillno_view_dmsServer()
}
