% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOut.R
\name{saleOut_sync_erpServer}
\alias{saleOut_sync_erpServer}
\title{销售出库同步至ERP}
\usage{
saleOut_sync_erpServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
销售出库同步至ERP
}
\examples{
saleOut_sync_erpServer()
}
