% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleReturn.R
\name{saleReturn_fbillno_view_dmsServer}
\alias{saleReturn_fbillno_view_dmsServer}
\title{销售退货按单据编号查询}
\usage{
saleReturn_fbillno_view_dmsServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
销售退货按单据编号查询
}
\examples{
saleReturn_fbillno_view_dmsServer()
}
