% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleReturn.R
\name{saleReturn_sync_erpServer}
\alias{saleReturn_sync_erpServer}
\title{销售退货同步至ERP}
\usage{
saleReturn_sync_erpServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
销售退货同步至ERP
}
\examples{
saleReturn_sync_erpServer()
}
