% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_miscellaneous_sal.R
\name{stk_miscellaneous_sal_fbillno_view_dmsServer}
\alias{stk_miscellaneous_sal_fbillno_view_dmsServer}
\title{DMS其他入库-工程不良按单据编号查询}
\usage{
stk_miscellaneous_sal_fbillno_view_dmsServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
DMS其他入库-工程不良按单据编号查询
}
\examples{
stk_miscellaneous_sal_fbillno_view_dmsServer()
}
