% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_miscellaneous_sal.R
\name{stk_miscellaneous_sal_sync_dmsServer}
\alias{stk_miscellaneous_sal_sync_dmsServer}
\title{其他入库-工程不良同步至数据中台}
\usage{
stk_miscellaneous_sal_sync_dmsServer(
  input,
  output,
  session,
  dms_token,
  wms_token
)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}

\item{wms_token}{}
}
\value{
返回值
}
\description{
其他入库-工程不良同步至数据中台
}
\examples{
stk_miscellaneous_sal_sync_dmsServer()
}
