% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_miscellaneous_std.R
\name{stk_miscellaneous_stdServer}
\alias{stk_miscellaneous_stdServer}
\title{处理逻辑}
\usage{
stk_miscellaneous_stdServer(input, output, session, dms_token, wms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
stk_miscellaneous_stdServer()
}
