% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_misdelivery_watste.R
\name{stk_misdelivery_watsteServer}
\alias{stk_misdelivery_watsteServer}
\title{处理逻辑}
\usage{
stk_misdelivery_watsteServer(input, output, session, dms_token, wms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
stk_misdelivery_watsteServer()
}
