% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_stktransferin_std.R
\name{stk_stktransferin_std_fbillno_view_dmsServer}
\alias{stk_stktransferin_std_fbillno_view_dmsServer}
\title{DMS标准直接调拨按单据编号查询}
\usage{
stk_stktransferin_std_fbillno_view_dmsServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
DMS标准直接调拨按单据编号查询
}
\examples{
stk_stktransferin_std_fbillno_view_dmsServer()
}
