% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_stktransferin_std.R
\name{stk_stktransferin_std_sync_dmsServer}
\alias{stk_stktransferin_std_sync_dmsServer}
\title{标准直接调拨同步至数据中台}
\usage{
stk_stktransferin_std_sync_dmsServer(
  input,
  output,
  session,
  dms_token,
  wms_token
)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}

\item{wms_token}{}
}
\value{
返回值
}
\description{
标准直接调拨同步至数据中台
}
\examples{
stk_stktransferin_std_sync_dmsServer()
}
