% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_stktransferin_std.R
\name{stk_stktransferin_std_sync_erpServer}
\alias{stk_stktransferin_std_sync_erpServer}
\title{标准直接调拨同步至ERP}
\usage{
stk_stktransferin_std_sync_erpServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
标准直接调拨同步至ERP
}
\examples{
stk_stktransferin_std_sync_erpServer()
}
